Address remaining review feedback
- Use ParseIssueFilterStateIsClosed for ListProjects state parsing
- Add SortTypeProjectColumnSorting const, replace magic string
- Use GetIssueByRepoID and dedupe Add/Remove issue handlers
- Migrate Column.Sorting from int8 to int (drops 127-column limit, allows
  the API to expose a normal int without truncation)
- Introduce project_service.UpdateProject with optional.Option fields,
  use it from the API EditProject handler

+++ services/forms/repo_form.go
@@ -469,7 +469,7 @@ type CreateProjectForm struct {
 // EditProjectColumnForm is a form for editing a project column
 type EditProjectColumnForm struct {
 	Title   string `binding:"Required;MaxSize(100)"`
-	Sorting int8
+	Sorting int
 	Color   string `binding:"MaxSize(7)"`
 }
